refactor: migrate home page components to new landing directory structure
