+++ apps/frontend/src/app/app.html
@@ -2,5 +2,7 @@
   class="min-h-screen transition-colors duration-300 bg-white dark:bg-gray-900"
   [class.dark]="themeService.isDarkMode()"
 >
+  <app-header />
   <router-outlet />
+  <app-footer />
 </div>
