+++ apps/frontend/src/app/pages/home/home.component.ts
@@ -1,36 +1,7 @@
-import { Component, signal } from '@angular/core';
-import { HeaderComponent } from './components/header/header';
-import { CtaComponent } from './components/cta/cta';
-import { AboutComponent } from './components/about/about';
-import { TestimonialsComponent } from './components/testimonials/testimonials';
-import { FeaturesComponent } from './components/features/features';
-import { FooterComponent } from './components/footer/footer';
-import { ThemeButtonComponent } from './components/theme-button/theme-button';
-import { testimonials, aboutDetails, features } from './home-constants';
+import { Component } from "@angular/core";
 
 @Component({
-  selector: 'app-home',
-  imports: [
-    HeaderComponent,
-    CtaComponent,
-    AboutComponent,
-    TestimonialsComponent,
-    FeaturesComponent,
-    FooterComponent,
-    ThemeButtonComponent,
-  ],
-  template: `
-    <app-theme-button />
-    <app-header />
-    <app-about [aboutDetails]="aboutDetails()" />
-    <app-features [features]="features()" />
-    <app-testimonials [testimonials]="testimonials()" />
-    <app-cta />
-    <app-footer />
-  `,
+  selector: "app-home",
+  template: ``,
 })
-export default class HomeComponent {
-  testimonials = signal(testimonials);
-  aboutDetails = signal(aboutDetails);
-  features = signal(features);
-}
+export class HomeComponent { }
