+++ apps/frontend/src/app/pages/landing/components/about/about.ts
