+++ apps/frontend/src/app/pages/landing/components/features/features.ts
