+++ apps/frontend/src/app/pages/landing/landing-constants.ts
