+++ apps/frontend/src/app/pages/landing/types.ts
